pyproject.toml
[tool.poetry]
name = "circuit-analysis-api"
version = "0.1.0"
description = "API for analyzing electronic circuit components using computer vision"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
uvicorn = "^0.27.1"
pydantic = "^2.6.3" 
pydantic-settings = "^2.2.1"
requests = "^2.31.0"
opencv-python = "^4.9.0.80"
loguru = "^0.7.2"
litellm = "^1.52.15"
arize-phoenix = "^5.12.0"
arize-phoenix-otel = "^0.6.1"
groq = "^0.9.0"
openinference-instrumentation-litellm = "^0.1.5"
openinference-instrumentation-groq = "^0.1.4"
openinference-instrumentation-openai = "^0.1.18"
dynaconf = "^3.2.5"
Pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
